Update com/dao/SkuDAO.cfc
A slight correction was needed to the cast to bigint. We were still running into the arithmetic overflow error on occasion. We were able to correct by casting the individual numbers within the power function so that all numbers were already typed as bigint before the power function even ran. 